backend/requirements.txt
aiohttp==3.9.5
db==0.1.1
fastapi==0.115.11
llama_index==0.12.24
nest_asyncio==1.6.0
openai==1.66.3
pydantic==2.10.6
python-dotenv==1.0.1
Requests==2.32.3
uvicorn==0.34.0
